Cap4 - Adicionando Update e Delete nos serviços de leilão e Search

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -91,6 +91,8 @@ namespace AuctionService.Controllers
             auction.Item.Mileage = updateAuctionDto.Mileage ?? auction.Item.Mileage;
             auction.Item.Year = updateAuctionDto.Year ?? auction.Item.Year;
 
+            await publishEndpoint.Publish(_mapper.Map<AuctionUpdated>(auction));
+
             var result = await _context.SaveChangesAsync() > 0;
 
             if (result) return Ok();
@@ -103,9 +105,13 @@ namespace AuctionService.Controllers
         {
             var auction = await _context.Auctions.FindAsync(id);
 
+            if (auction == null) return NotFound();
+
             //TODO - check seller == username
 
             _context.Auctions.Remove(auction);
+
+            await publishEndpoint.Publish<AuctionDeleted>(new { Id = auction.Id.ToString() });
 
             var result = await _context.SaveChangesAsync() > 0;
